+++ tests/testFunctions.sh
@@ -2,6 +2,6 @@
 # isSameString $EXPECTED $ACTUAL
 # Returns 0 on success 1 on failure.
 isSameString () {
-    [[ "$1" == "$2" ]] && return 0 || return 1
+    [[ "$1" == "$2" ]]
 }